Rewrite and improvement of current classes...
New packages:
	I added the 'newCTF' package, which is going to replace the current ctf
package. It slightly changes the layout of Team classes, so I made a new
sample team to use as an example and to test with.

Changes to ctf:
	The new newCTF package has all of the same functionality as the old
one, but I re-wrote and re-structured all of the classes to make them
more intuitive. I also fixed a surprising number of bugs. The classes
generally do what they sound like. A match runs a game between two
teams, a tournament organizes a series of matches, and a bracket
displays the bracket of a tournament(currently broken). I extended
ActorWorld so now theoretically it should work of the default GridWorld
packages. The tournaments currently run all of the bottom games before
moving to the next level.

Changes to come:
	I need to make Bracket work, but currently I'm not very good with
Jframes.
	I need to make Match or CtfWorld keep track of and announce stats about
the game.
	I need to extend more GridWorld classes to fix the zooming, and add
other ease of use features like an end game button

	-Connor